Add 'premier' image asset and update related constants for new project 'Premier Model Management'
This commit adds a new image asset 'premier.png' to the assets folder and updates the related constants in the index.js file. It also adds a new project object for 'Premier Model Management' in the projects array in the index.js file under the constants folder, utilizing the new image asset and constants.